encourager/requirements.txt
APScheduler==3.10.4
backports.zoneinfo==0.2.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
Werkzeug==2.3.4
zipp==3.15.0
